Add a function getProjectPath
This function is implemented in Hitweb.Filepath.
It check if a project is a repo or contains a repo
(is the project is a .git directory or if it contains a .git directory)

It is useful to do this test because on server, we do not generate the project
but only keep a bare git (the .git directory).